HW11_End_to_End_Learning/task1/requirements.txt
numpy==1.20.1
pandas==1.2.3
torch==1.9.0
torchvision==0.10.0
matplotlib==3.3.4
Pillow==8.2.0
pytest==6.2.2
